requirements.txt
langchain
langchain-openai
python-dotenv
openpyxl
python-docx
PyPDF2
tqdm
